Move eth api trait and impl into reth-rpc-eth-api
