+++ crates/rpc/rpc-api/Cargo.toml
@@ -29,3 +29,4 @@ client = [
     "jsonrpsee/async-client",
     "reth-rpc-eth-api/client"
 ]
+optimism = ["reth-rpc-eth-api/optimism"]
