+++ crates/rpc/rpc-eth-api/src/eth/api/traits/mod.rs
@@ -11,7 +11,7 @@
 //! divided into: [`EthTransactions`], [`EthBlocks`], [`EthFees`], [`EthState`] and [`EthCall`].
 //! Default implementation of the `Eth` traits, is done w.r.t. L1.
 //!
-//! [`EthApiServer`](reth_rpc_api::EthApiServer), is implemented for any type that implements all
+//! [`EthApiServer`](reth_rpc_eth_api::EthApiServer), is implemented for any type that implements all
 //! the `Eth` traits, e.g. [`EthApi`](crate::EthApi).
 
 pub mod block;
