+++ crates/rpc/rpc-eth-api/src/eth/api/traits/transaction.rs
@@ -29,7 +29,7 @@ use crate::eth::{
     TransactionSource,
 };
 
-/// Transaction related functions for the [`EthApiServer`](reth_rpc_api::EthApiServer) trait in
+/// Transaction related functions for the [`EthApiServer`](reth_rpc_eth_api::EthApiServer) trait in
 /// the `eth_` namespace.
 ///
 /// This includes utilities for transaction tracing, transacting and inspection.
