+++ crates/rpc/rpc-eth-api/src/eth/utils.rs
@@ -6,7 +6,7 @@ use reth_primitives::{Bytes, PooledTransactionsElement, PooledTransactionsElemen
 /// Recovers a [`PooledTransactionsElementEcRecovered`] from an enveloped encoded byte stream.
 ///
 /// See [`PooledTransactionsElement::decode_enveloped`]
-pub(crate) fn recover_raw_transaction(
+pub fn recover_raw_transaction(
     data: Bytes,
 ) -> EthResult<PooledTransactionsElementEcRecovered> {
     if data.is_empty() {
