+++ crates/rpc/rpc-testing-util/Cargo.toml
@@ -16,6 +16,7 @@ workspace = true
 reth-primitives.workspace = true
 reth-rpc-types.workspace = true
 reth-rpc-api = { workspace = true, features = ["client"] }
+reth-rpc-eth-api = { workspace = true, features = ["client"] }
 
 # async
 futures.workspace = true
