+++ crates/rpc/rpc-testing-util/tests/it/trace.rs
@@ -1,6 +1,6 @@
 use futures::StreamExt;
 use jsonrpsee::http_client::HttpClientBuilder;
-use reth_rpc_api::EthApiClient;
+use reth_rpc_eth_api::EthApiClient;
 use reth_rpc_api_testing_util::{debug::DebugApiExt, trace::TraceApiExt, utils::parse_env_url};
 use reth_rpc_types::trace::{
     filter::TraceFilter, parity::TraceType, tracerequest::TraceCallRequest,
