+++ crates/rpc/rpc/src/net.rs
@@ -1,8 +1,8 @@
-use crate::eth::EthApiSpec;
 use jsonrpsee::core::RpcResult as Result;
 use reth_network_api::PeersInfo;
 use reth_primitives::U64;
 use reth_rpc_api::NetApiServer;
+use reth_rpc_eth_api::eth::EthApiSpec;
 use reth_rpc_types::PeerCount;
 
 /// `Net` API implementation.
