+++ crates/rpc/rpc/src/otterscan.rs
@@ -3,6 +3,10 @@ use async_trait::async_trait;
 use jsonrpsee::core::RpcResult;
 use reth_primitives::{Address, BlockId, BlockNumberOrTag, TxHash, B256};
 use reth_rpc_api::{EthApiServer, OtterscanServer};
+use reth_rpc_eth_api::{
+    eth::api::{LoadBlock, LoadTransaction, TraceExt},
+    result::internal_rpc_err,
+};
 use reth_rpc_types::{
     trace::otterscan::{
         BlockDetails, ContractCreator, InternalOperation, OperationType, OtsBlockTransactions,
@@ -12,11 +16,6 @@ use reth_rpc_types::{
 };
 use revm_inspectors::transfer::{TransferInspector, TransferKind};
 use revm_primitives::ExecutionResult;
-
-use crate::{
-    eth::api::{LoadBlock, LoadTransaction, TraceExt},
-    result::internal_rpc_err,
-};
 
 const API_LEVEL: u64 = 8;
