+++ crates/rpc/rpc/src/trace.rs
@@ -7,6 +7,12 @@ use reth_primitives::{revm::env::tx_env_with_recovered, BlockId, Bytes, SealedHe
 use reth_provider::{BlockReader, ChainSpecProvider, EvmEnvProvider, StateProviderFactory};
 use reth_revm::database::StateProviderDatabase;
 use reth_rpc_api::TraceApiServer;
+use reth_rpc_eth_api::eth::{
+    api::{LoadBlock, LoadTransaction, TraceExt},
+    error::{EthApiError, EthResult},
+    revm_utils::{prepare_call_env, EvmOverrides},
+    utils::recover_raw_transaction,
+};
 use reth_rpc_types::{
     state::StateOverride,
     trace::{
@@ -27,13 +33,6 @@ use revm_inspectors::{
     tracing::{parity::populate_state_diff, TracingInspector, TracingInspectorConfig},
 };
 use tokio::sync::{AcquireError, OwnedSemaphorePermit};
-
-use crate::eth::{
-    api::{LoadBlock, LoadTransaction, TraceExt},
-    error::{EthApiError, EthResult},
-    revm_utils::{prepare_call_env, EvmOverrides},
-    utils::recover_raw_transaction,
-};
 
 /// `trace` API implementation.
 ///
